oprofile/mips: do not set perf_irq for qemu mips 64
In a similar manner to the following commit:

    oprofile/mips: override register writes for qemu mips 64

    Even when oprofile timer mode is enabled, during oprofile arch init
    there are writes to registers that do not exist in the qemu malta
    64 bit pseudo machine.

    To allow initialization to continue and the fallback to oprofile
    timer mode to take effect, we stub the register writes.

    These two defines are only appropriate for qemu-system-mips.
    This should not merge to a common location, and only be applied
    to emulated boards.

We also cannot set perf_irq to mipsxx_perfcount_handler during oprofile
arch init, or unsupported register reads will be triggered before the
fallback to timer mode occurs. This commit is also only appropriate for
qemu mips based boards, and should not be applied to any general branch.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>