mtd: rawnand: Add driver for Arasan Nand Controller
Added basic version of driver for Arasan Nand Controller.
Added entry for arasan_nfc driver in Kconfig and Makefile.

Do not increment the fail count in anfc_read_page_hwecc(),as we don't
have mechanism to read ecc errors in arasan nand controller, if the
errors are more than 24-bit.
This update is there in older arasan nand driver but not in the 4.19
arasan driver.

Remove setup_data_interface() hook, as it is supporting only
SDR modes 0 to 5. Because of this NV-DDR support was broken in the
latest driver. so add back this support which was there in older
version of driver.

Signed-off-by: Shakti Bhatnagar <shaktib@xilinx.com>
Signed-off-by: Naga Sureshkumar Relli <naga.sureshkumar.relli@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
[Quanyang: original patch taken from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v4.19]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>