ARM: Activate FIQs to avoid __irq_startup warnings
commit  64fd4317605929f35a82381dda1cd56b77b0b685 from
https://github.com/raspberrypi/linux.git rpi-5.12.y

There is a new test in __irq_startup that the IRQ is activated, which
hasn't been the case for FIQs since they bypass some of the usual setup.

Augment enable_fiq to include a call to irq_activate to avoid the
warning.

Signed-off-by: Phil Elwell <phil@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>