spi: spi-zynqmp-gqspi: use dma mode only if the buffer is not vmalloced
As per the kernel documentation the buffer to be used with dma
should not be vmalloced due to the fact that if buffer is vmalloced
then page entries are not consistent in physical pages. That would
lead to failure cases in dma. This patch adds condition to check
if the buffer to be read is vmalloced or not if not vmalloced then
uses dma mode otherwise io mode.

Signed-off-by: Tejas Prajapati Rameshchandra <tejas.prajapati.rameshchandra@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
[Quanyang: original patch taken from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v4.19]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>