of: configfs: Use of_overlay_fdt_apply API call
commit  8203d9314cfcde04bf7837fd1690dd521c8017da from
https://github.com/raspberrypi/linux.git rpi-5.10.y

The published API to the dynamic overlay application mechanism now
takes a Flattened Device Tree blob as input so that it can manage the
lifetime of the unflattened tree. Conveniently, the new API call -
of_overlay_fdt_apply - is virtually a drop-in replacement for
create_overlay, which can now be deleted.

Signed-off-by: Phil Elwell <phil@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>