bcm2835-dma: Add support for per-channel flags
commit  afc57a94dc7833b282f54d64c6c2197d0779acf6 from
https://github.com/raspberrypi/linux.git rpi-5.11.y

Add the ability to interpret the high bits of the dreq specifier as
flags to be included in the DMA_CS register. The motivation for this
change is the ability to set the DISDEBUG flag for SD card transfers
to avoid corruption when using the VPU debugger.

Signed-off-by: Phil Elwell <phil@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>