lan78xx: Debounce link events to minimize poll storm
commit  84605f467f98769901031a263f834829098dc650 from
https://github.com/raspberrypi/linux.git rpi-5.11.y

The bInterval is set to 4 (i.e. 8 microframes => 1ms) and the only bit
that the driver pays attention to is "link was reset". If there's a
flapping status bit in that endpoint data, (such as if PHY negotiation
needs a few tries to get a stable link) then polling at a slower rate
would act as a de-bounce.

See: https://github.com/raspberrypi/linux/issues/2447
Signed-off-by: Meng Li <Meng.Li@windriver.com>