drm: xlnx: dsi: Fix kernel panic when "halt" is run
When "halt" is run, the kernel panics and dumps a kernel call stack
which showing it originating in xlnx_dsi_unbind(). On closer
examination, it is found that the xlnx_dsi_host_detach() incorrectly
sets the "panel_mode" device node member pointer to NULL instead of
setting the drm_panel member pointer "panel" to NULL.

The patch corrects this incorrect assignment.

Signed-off-by: Vishal Sagar <vishal.sagar@xilinx.com>
Reviewed-by: Sandip Kothari <sandipk@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
[Quanyang: original patch taken from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v4.19]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>