staging: mmal-vchiq: Fix client_component for 64 bit kernel
commit  9994fe6a1ada549f787d016e7aa7213eceaee06e from
https://github.com/raspberrypi/linux.git rpi-5.12.y

The MMAL client_component field is used with the event
mechanism to allow the client to identify the component for
which the event is generated.
The field is only 32bits in size, therefore we can't use a
pointer to the component in a 64 bit kernel.

Component handles are already held in an array per VCHI
instance, so use the array index as the client_component handle
to avoid having to create a new IDR for this purpose.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>