staging: mmal-vchiq: If the VPU returns an error, don't negate it
commit  e6623644777733344471f36eaaad158ff1128856 from
https://github.com/raspberrypi/linux.git rpi-5.5.y

There is an enum for the errors that the VPU can return.
port_parameter_get was negating that value, but also using -EINVAL
from the Linux error codes.
Pass the VPU error code as positive values. Should the function
need to pass a Linux failure, then return that as negative.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>