PCI: brcmstb: Add MSI capability
commit  e5c5c76b83b44cab32a75c7ef789d4049ce2bc99 from
https://github.com/raspberrypi/linux.git rpi-5.5.y

This commit adds MSI to the Broadcom STB PCIe host controller. It does
not add MSIX since that functionality is not in the HW.  The MSI
controller is physically located within the PCIe block, however, there
is no reason why the MSI controller could not be moved elsewhere in
the future.

Since the internal Brcmstb MSI controller is intertwined with the PCIe
controller, it is not its own platform device but rather part of the
PCIe platform device.

Signed-off-by: Jim Quinlan <jim2101024@gmail.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>