PCI: brcmstb: Add Broadcom STB PCIe host controller driver
commit  0741358f97df77f25f58cdfb04354d96450c3c44 from
https://github.com/raspberrypi/linux.git rpi-5.5.y

This commit adds the basic Broadcom STB PCIe controller.  Missing is
the ability to process MSI and also handle dma-ranges for inbound
memory accesses.  These two functionalities are added in subsequent
commits.

The PCIe block contains an MDIO interface.  This is a local interface
only accessible by the PCIe controller.  It cannot be used or shared
by any other HW.  As such, the small amount of code for this
controller is included in this driver as there is little upside to put
it elsewhere.

Signed-off-by: Jim Quinlan <jim2101024@gmail.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>