usb: xhci: Disable the XHCI 5 second timeout
commit  339cf297e212f1a1fba8720c8f423537e0e805cb from
https://github.com/raspberrypi/linux.git rpi-5.11.y

If the VL805 EEPROM has not been programmed then boot will hang for five
seconds. The timeout seems to be arbitrary and is an unecessary
delay on the first boot. Remove the timeout.

This is common code and probably can't be upstreamed unless the timeout
can be configurable somehow or perhaps the XHCI driver can be skipped
on the first boot.
Signed-off-by: Meng Li <Meng.Li@windriver.com>