usb: xhci: Disable the XHCI 5 second timeout
commit  578fadc9e54fa0b2c71a00ee27521633c3875581 from
https://github.com/raspberrypi/linux.git rpi-5.5.y

If the VL805 EEPROM has not been programmed then boot will hang for five
seconds. The timeout seems to be arbitrary and is an unecessary
delay on the first boot. Remove the timeout.

This is common code and probably can't be upstreamed unless the timeout
can be configurable somehow or perhaps the XHCI driver can be skipped
on the first boot.
Signed-off-by: Meng Li <Meng.Li@windriver.com>