v4l: xilinx: xcsi2rxss: Use external reset in a SLBF condition
The IP document recommends to use external reset, video_resetn, for a
stream line buffer full condition. Without this re-running the
pipeline will continue to result in stream line buffer full condition
as the FIFO is not cleared with IP soft reset.

So add external reset support using gpio and use it as optional
feature.

Signed-off-by: Subhransu S. Prusty <subhransu.sekhar.prusty@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
[Quanyang: original patch taken from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v4.19]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>