drm/vc4: Resolve the vblank warnings on mode switching
commit  84f5888f14fa0c8c8f4422b66dbd5bb510661745 from
https://github.com/raspberrypi/linux.git rpi-5.5.y

The details over when and how a driver is to service the
vblank events are sketchy, and the fkms driver was triggering
a kernel warning every time the crtc was enabled or disabled.

Copy the event handling as used by the vc4-kms driver slightly
more closely, and we avoid the warnings.

https://github.com/raspberrypi/linux/issues/3020

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>