pcie-brcmstb: Don't set DMA ops for root complex
commit  68770cb42030020b0d358bb44444737bfafc1939 from
https://github.com/raspberrypi/linux.git rpi-5.5.y

A change to arm_get_dma_map_ops has stopped get_dma_ops from working
on the root complex, causing an error to be logged. However, there is
no need to override the DMA ops in that case, so skip it and
eliminate the error message.

Signed-off-by: Phil Elwell <phil@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>