mmc: cavium-thunderx: Drop the IRQF_NO_THREAD constraint
The IRQF_NO_THREAD is added by a Marvell SDK patch 238e623e1024 ("mmc:
cavium: fix swiotlb buffer is full") in order to get back some of the
performance loss. But in some cases (such as rt kernel), we do need the
capability to thread irq handler. Otherwise we would get warnings because
the normal spin lock is used in the irq handler. So drop this constraint.

Signed-off-by: Kevin Hao <haokexin@gmail.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>