staging: vchiq_arm: Give vchiq children DT nodes
commit  304df4143f970e83cba5413a87f4321999931037 from
https://github.com/raspberrypi/linux.git rpi-5.12.y

vchiq kernel clients are now instantiated as platform drivers rather
than using DT, but the children of the vchiq interface may still
benefit from access to DT properties. Give them the option of a
a sub-node of the vchiq parent for configuration and to allow
them to be disabled.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>