bpf, lpm: Make locking RT friendly
The LPM trie map cannot be used in context like perf, kprobes and tracing
as they dynamically allocate memory. The memory allocation happens with a
raw spinlock held which is a truly spinning lock on a PREEMPT RT enabled
kernel which disables preemption and interrupts.

As RT does not allow memory allocation from such a section for various
reasons, convert the raw spinlocks to regular spinlocks. On a RT enabled
kernel these locks are substituted by 'sleeping' spinlocks which provide
the proper protection but keep the code preemptible.

On a non-RT kernel regular spinlocks map to raw spinlocks, i.e. this does
not cause any functional change.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>