mm: Revert the DEFINE_PER_CPU_PAGEVEC implementation
In the v5.4-RT cycle the local-lock based locking for struct pagevec was
replaced with explicit per-CPU locking of the struct pagevec. This
explicit locking was used if NO_HZ_FULL was active and if was always
used on -RT.
Upstream wasn't too excited about this so we will probably stick with
the local-lock based "fix" to deal with the required locking.
This is an all-in-one change containing
- a revert of the following patches:
   0001-mm-page_alloc-Split-drain_local_pages.patch
   0002-mm-swap-Add-static-key-dependent-pagevec-locking.patch
   0003-mm-swap-Access-struct-pagevec-remotely.patch
   0004-mm-swap-Enable-use_pvec_lock-nohz_full-dependent.patch
   mm-swap-Enable-use-pvec-lock-on-RT.patch

- adding back the old patches:
  mm-convert-swap-to-percpu-locked.patch
  mm-perform-lru_add_drain_all-remotely.patch

Since the v5.4 series will be maintained for a bit longer I don't want
to divert the trees more than they have to and so this brings the old
behaviour we had in v5.2-RT and earlier.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>