sdhci: arasan: use an array of integers for *_itap/otap_delays instead of macro *_ITAP/OTAP_DELAYS
When assigning value to a pointer as below:

	itapdly = (u32 [MMC_TIMING_MMC_HS400 + 1]) {0x0, 0x15, 0x15...0x0};

then using gcc 9.2.0 to compile, in some condition the pointer itapdly
will point to the stack of the function, and the values in this area is
not initialized. The disassemble code is as below:

	ffffffc010a389b8:   940089ac    bl  ffffffc010a5b068 <of_device_is_compatible>
	ffffffc010a389bc:   34001100    cbz w0, ffffffc010a38bdc <arasan_dt_parse_tap_delays+0x264>
	        itapdly = (u32 [MMC_TIMING_MMC_HS400 + 1]) ZYNQMP_ITAP_DELAYS;
	        otapdly = (u32 [MMC_TIMING_MMC_HS400 + 1]) ZYNQMP_OTAP_DELAYS;
	ffffffc010a389c0:   9101d3f4    add x20, sp, #0x74
	        itapdly = (u32 [MMC_TIMING_MMC_HS400 + 1]) ZYNQMP_ITAP_DELAYS;
	ffffffc010a389c4:   910123f5    add x21, sp, #0x48
	    int ret = of_property_read_variable_u32_array(np, propname, out_values,

This results that the pointer itapdly is not the address of the array
ZYNQMP_ITAP_DELAYS but a stack address which contains random values.

So use an array of integers to avoid the over optimization of the new
compiler.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>