media: bcm2835-unicam: Use dummy buffer if none have been queued
commit  9e41ab9bec7da7453af00cf46e4505959ec0001c from
https://github.com/raspberrypi/linux.git rpi-5.10.y

If no buffer has been queued by a userland application, we use an
internal dummy buffer for the hardware to spin in. This will allow
the driver to release the existing userland buffer back to the
application for processing.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>