media: bcm2835-unicam: Use dummy buffer if none have been queued
commit  bebc4d3f3e254b7b6fee6585288792efeed3950f from
https://github.com/raspberrypi/linux.git rpi-5.12.y

If no buffer has been queued by a userland application, we use an
internal dummy buffer for the hardware to spin in. This will allow
the driver to release the existing userland buffer back to the
application for processing.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>