vc4: Report channel mapping back to userspace
commit  8f8d33a98ab77e1d1b60be0300797daadd22595a from
https://github.com/raspberrypi/linux.git rpi-5.10.y

This follows logic in hdmi-codec.c to use speaker layout
from ELD to choose a suitable speaker mapping based on
number of channels requested and signal that in audio
infoframe  and report this back to userspace.

This allows apps like speaker-test and kodi to get the
output to the right speakers.

Signed-off-by: Dom Cobley <popcornmix@gmail.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>