vc4: Report channel mapping back to userspace
commit  98f804f9060f2b5c343bccb32db4cfe0620b90cc from
https://github.com/raspberrypi/linux.git rpi-5.12.y

This follows logic in hdmi-codec.c to use speaker layout
from ELD to choose a suitable speaker mapping based on
number of channels requested and signal that in audio
infoframe  and report this back to userspace.

This allows apps like speaker-test and kodi to get the
output to the right speakers.

Signed-off-by: Dom Cobley <popcornmix@gmail.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>