media: bcm2835-unicam: Do not stop streaming in unicam_release
commit  e503d07ceb20589df80dd025277f13884ac4c3f5 from
https://github.com/raspberrypi/linux.git rpi-5.12.y

unicam_release calls _vb2_fop_release, which will call stop_streaming
if that particular node was streaming. Calling it unconditionally (as
the code was) means that if a second handle was opened eg to alter
a setting, on closing that connection it also stopped Unicam.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>