xilinx: Disable strict devmem access control
commit  22b71b41620dac13c69267d2b7898ebfb14c954e from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.4

Intention of the common kernel is ease of use and "demonstration of
capabilities". The intention of these defconfigs is not to deploy these
configurations on any real product.
With keeping this in mind the patch disable strict devmem access control
which can be used by some user space applications.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: not-upstreamable
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>