configs: Make the evdev module a built-in
commit  23093cdadb90b558a5e82d59436c69d589c91d8b from
https://github.com/raspberrypi/linux.git rpi-5.9.y

The evdev module is required to use a keyboard (although not a mouse,
it appears). This means that if a kernel is unable to load modules the
user will be unable to interact with it by attaching a keyboard.

For the cost of a few KBs in kernel bloat, make evdev a built-in
driver.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>