octeontx2-sdp: Changed sdp dts node read code.
commit 6c93969fe29cc81b538535fdfd5696ca9136af43 from
git@git.assembla.com:cavium/WindRiver.linux.git

The ring distribution between SDP PF and its
corresponding VFs is not specified in the dts. Sdp
dts node is expanded to have details such as rings
available for PF and ring limits for each of the
VFs as an array. In case there are multiple SDP
PFs, the starting ring number for a PF is also
mentioned in the dts node from which that
particular PF can have "num_pf_rings". As dts sdp
node contents changed, the code that parses the
same is changed accordingly.

Below is the field description:
num_pf_rings: Rings available for SDP PF use.
pf_srn: Starting ring for  this PF.
num_vf_rings: An array of ring counts for each of
the VFs. If ring count is not mentioned for a VF,
the ring count of previous VF is considered.

Change-Id: I4a55b7f5575b9ab0bb60ffef1efc9dad4883b9e8
Signed-off-by: Nalla, Pradeep <pnalla@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/35222
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Satananda Burla <sburla@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Signed-off-by: Abhijit Ayarekar <aayarekar@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/49035
Reviewed-by: Radha Chintakuntla <radhac@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>