overlays: Update i2c0 overlay to disable the i2c0mux.
commit  88464390de29cc9b4ea9b57748b12067c849fe16 from
https://github.com/raspberrypi/linux.git rpi-5.9.y

The i2c0 overlay was assigning pinctrl settings to node i2c0,
which is now the port@0 output of the mux. That leaves a high
chance of it colliding with the port@1 output and not doing
what was intended.

Set the i2c0 overlay to disable i2c0mux, set the pin-ctrl on
the root i2c controller, and redirect the alias, so overall it
behaves exactly as before.
Combining with dtparam=i2c_vc=on is going to cause conflicts,
so this is noted in the README.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>