drm/vc4: hdmi: Create a custom connector state
commit  cd8f7ddafda853ebf01cf47efa159c9b062591f3 from
https://github.com/raspberrypi/linux.git rpi-5.10.y

When run with a higher bpc than 8, the clock of the HDMI controller needs
to be adjusted. Let's create a connector state that will be used at
atomic_check and atomic_enable to compute and store the clock rate
associated to the state.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>