drm/vc4: hdmi: Store pixel frequency in the connector state
commit  19b67358329015ec110b3519854a1029e3ee02ac from
https://github.com/raspberrypi/linux.git rpi-5.10.y

The pixel rate is for now quite simple to compute, but with more features
(30 and 36 bits output, YUV output, etc.) will depend on a bunch of
connectors properties.

Let's store the rate we have to run the pixel clock at in our custom
connector state, and compute it in atomic_check.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>