media: bcm2835-unicam: Return early from stop_streaming() if stopped
commit  e6349c763e5ec88f793ee441bcf13f6131a8576f from
https://github.com/raspberrypi/linux.git rpi-5.10.y

clk_disable_unprepare() is called unconditionally in stop_streaming().
This is incorrect in the cases where start_streaming() fails, and
unprepares all clocks as part of the failure cleanup. To avoid this,
ensure that clk_disable_unprepare() is only called in stop_streaming()
if the clocks are in a prepared state.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>