i2c: bcm2835: Handle untimely DONE signal
commit  a1efce8ad0fa6c9c6600ef238592890397c77fa1 from
https://github.com/raspberrypi/linux.git rpi-5.12.y

Under certain circumstance the DONE flag can appear to be set early.
Fortunately the TA flag is often still set at that time, and it can be
used as an indication that DONE is invalid.

Handle the other cases - when TA is not set but not all data is
available - by silently accepting it and hoping for a second DONE.

See: https://github.com/raspberrypi/linux/issues/3064

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>