gpio-fsm: Rename 'num-soft-gpios' to avoid warning
commit  454ed79452c6ea33a62ae4c9f34188027167dff8 from
https://github.com/raspberrypi/linux.git rpi-5.14.y

As of 5.10, the Device Tree parser warns about properties that look
like references to "suppliers" of various services. "num-soft-gpios"
resembles a declaration of a GPIO called "num-soft", causing the value
to be interpreted as a phandle, the owner of which is checked for a
"#gpio-cells" property.

To avoid this warning, rename the gpio-fsm property to "num-swgpios".

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>