net: thunderx: Modify BGX driver probe
commit 15d97aa96b20cd9a46a3ff9f5729481440cf9888 from
git@git.assembla.com:cavium/WindRiver.linux.git

current design is that, BGX driver probe fails in case
of failure in finding of phy device associated with
any one of lmac. Which results in all lmacs connected that
BGX wont be available to NIC driver.

This patch changes the probe of BGX device such that only
lmacs failed in discovering phy device will be released and
other lmacs will be available to NIC driver usage.

With this changes
	- Lmacs which are not associated with phy device wont be
	  enabled and bgx lmac count is updated accordingly

Change-Id: I43851391cbdc8dceeb140dd5e042f6a6e3227ec1
Signed-off-by: Hariprasad Kelam <hkelam@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/47229
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>