drm/vc4: Fix VEC address for BCM2711 in the devicetrees
commit  e45b90c951021757521210721865b14fa9331137 from
https://github.com/raspberrypi/linux.git rpi-5.12.y

The VEC has a different address (0x7ec13000) on the BCM2711 (used in
e.g. Raspberry Pi 4) compared to BCM238x (e.g. Pi 3 and earlier). This
was erroneously not taken account for.

Definition of the VEC in the devicetrees had to be moved from
bcm283x.dtsi to bcm2711.dtsi and bcm2835-common.dtsi to allow for this
differentiation.

Signed-off-by: Mateusz Kwiatkowski <kfyatek+publicgit@gmail.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>