videodev2.h: add V4L2_CTRL_FLAG_DYNAMIC_ARRAY
commit  b95f7699069b3947416960feaa22d644816663d7 from
https://github.com/raspberrypi/linux.git rpi-5.14.y

Add a new flag that indicates that this control is a dynamically sized
array. Also document this flag.

Currently dynamically sized arrays are limited to one dimensional arrays,
but that might change in the future if there is a need for it.

The initial use-case of dynamic arrays are stateless codecs. A frame
can be divided in many slices, so you want to provide an array containing
slice information for each slice. Typically the number of slices is small,
but the standard allow for hundreds or thousands of slices. Dynamic arrays
are a good solution since sizing the array for the worst case would waste
substantial amounts of memory.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
(cherry picked from commit c095ad310b223b9de38d491c9f66533d05586b45)
Signed-off-by: Meng Li <Meng.Li@windriver.com>