overlays: Add ugreen-dabboard overlay
commit  302a3e0100a5c6db1c94c22e2d3f6374a4bd1a40 from
https://github.com/raspberrypi/linux.git rpi-5.12.y

This is a simple overlay based on the simple-audio-card and the dmic
codec. It has the speciality that it is configured to use the codec
as a master I2S device. It works for example with the Si468x DAB
receiver on the uGreen DABBoard.

See: https://github.com/raspberrypi/linux/issues/4304

Signed-off-by: Christoph Orth <c.orth@ugreen.eu>
Signed-off-by: Meng Li <Meng.Li@windriver.com>