drm/vc4: hdmi: Rely on interrupts to handle hotplug
commit  4f9be892e78195f11764bd03775680879f6583cd from
https://github.com/raspberrypi/linux.git rpi-5.12.y

DRM currently polls for the HDMI connector status every 10s, which can
be an issue when we connect/disconnect a display quickly or the device
on the other end only issues a hotplug pulse (for example on EDID
change).

Switch the driver to rely on the internal controller logic for the
BCM2711/RPi4.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>