drm/vc4: crtc: Lookup the encoder from the register at boot
commit  fc87271a5b453b0eee03738fde4b6f6490d14f29 from
https://github.com/raspberrypi/linux.git rpi-5.12.y

At boot, we can't rely on the vc4_get_crtc_encoder since we don't have a
state yet and thus will not be able to figure out which connector is
attached to our CRTC.

However, we have a muxing bit in the CRTC register we can use to get the
encoder currently connected to the pixelvalve. We can thus read that
register, lookup the associated register through the vc4_pv_data
structure, and then pass it to vc4_crtc_disable so that we can perform
the proper operations.

Fixes: 875a4d536842 ("drm/vc4: drv: Disable the CRTC at boot time")
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>