media: rpivid: Fix H265 aux ent reuse of the same slot
commit  f5b531914410ca6f41ad17862b3d8c73c7a0a416 from
https://github.com/raspberrypi/linux.git rpi-5.14.y

It is legitimate, though unusual, for an aux ent associated with a slot
to be selected in phase 0 before a previous selection has been used and
released in phase 2. Fix such that if the slot is found to be in use
that the aux ent associated with it is reused rather than an new aux
ent being created. This fixes a problem where when the first aux ent
was released the second was lost track of.

This bug spotted in Nick's testing. It may explain some other occasional,
unreliable decode error reports where dmesg included "Missing DPB AUX
ent" logging.

Signed-off-by: John Cox <jc@kynesim.co.uk>
Signed-off-by: Meng Li <Meng.Li@windriver.com>