drm/vc4: hdmi: Enable the scrambler on reconnection
commit  d97a52cd8d922b5df8f8f7001bd1d9ad9c2f294d from
https://github.com/raspberrypi/linux.git rpi-5.14.y

If we have a state already and disconnect/reconnect the display, the
SCDC messages won't be sent again since we didn't go through a disable /
enable cycle.

In order to fix this, let's call the vc4_hdmi_enable_scrambling function
in the detect callback if there is a mode and it needs the scrambler to
be enabled.

Fixes: 74465b84fa27 ("drm/vc4: hdmi: Enable the scrambler")
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>