drm/vc4: hdmi: Warn if we access the controller while disabled
commit  79b1f1cb16846278417923892ff18692c8ea41cc from
https://github.com/raspberrypi/linux.git rpi-5.14.y

We've had many silent hangs where the kernel would look like it just
stalled due to the access to one of the HDMI registers while the
controller was disabled.

Add a warning if we're about to do that so that it's at least not silent
anymore.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>